requirements.txt
aiosqlite==0.17.0
annotated-types==0.7.0
anyio==4.4.0
black==24.8.0
click==8.1.7
fastapi==0.115.0
h11==0.14.0
idna==3.10
iso8601==1.1.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pydantic==2.9.2
pydantic_core==2.23.4
pypika-tortoise==0.1.6
pytz==2024.2
sniffio==1.3.1
starlette==0.38.5
tortoise-orm==0.21.6
typing_extensions==4.12.2
uvicorn==0.30.6
